move things over to users file
