+++ internal/extsvc/azuredevops/testdata/golden/AzureServicesProfile.json
@@ -1,0 +1,7 @@
+{
+  "id": "473dec3e-03d7-6147-b106-0b0a7f766a92",
+  "displayName": "idan.varsano",
+  "emailAddress": "idan.varsano@sourcegraph.com",
+  "timestamp": "2023-01-20T19:20:51.6433333Z",
+  "publicAlias": "473dec3e-03d7-6147-b106-0b0a7f766a92"
+ }
