+++ internal/extsvc/azuredevops/testdata/vcr/AzureServicesProfile.yaml
@@ -1,0 +1,60 @@
+---
+version: 1
+interactions:
+- request:
+    body: ""
+    form: {}
+    headers: {}
+    url: https://app.vssps.visualstudio.com/_apis/profile/profiles/me?api-version=7.0
+    method: GET
+  response:
+    body: '{"displayName":"idan.varsano","publicAlias":"473dec3e-03d7-6147-b106-0b0a7f766a92","emailAddress":"idan.varsano@sourcegraph.com","coreRevision":411938451,"timeStamp":"2023-01-20T19:20:51.6433333+00:00","id":"473dec3e-03d7-6147-b106-0b0a7f766a92","revision":411938451}'
+    headers:
+      Access-Control-Expose-Headers:
+      - Request-Context
+      Activityid:
+      - 88152494-f8a1-4314-b2eb-4a210fb00d73
+      Cache-Control:
+      - no-cache, no-store, must-revalidate
+      Content-Type:
+      - application/json; charset=utf-8; api-version=7.0
+      Date:
+      - Wed, 01 Feb 2023 16:48:43 GMT
+      Etag:
+      - '"0"'
+      Expires:
+      - "-1"
+      Last-Modified:
+      - Fri, 20 Jan 2023 19:20:51 GMT
+      P3p:
+      - CP="CAO DSP COR ADMa DEV CONo TELo CUR PSA PSD TAI IVDo OUR SAMi BUS DEM NAV
+        STA UNI COM INT PHY ONL FIN PUR LOC CNT"
+      Pragma:
+      - no-cache
+      Request-Context:
+      - appId=cid-v1:20b3930f-73dc-453a-b660-e3891d782eef
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains
+      Vary:
+      - Accept-Encoding
+      X-Cache:
+      - CONFIG_NOCACHE
+      X-Content-Type-Options:
+      - nosniff
+      X-Frame-Options:
+      - SAMEORIGIN
+      X-Msedge-Ref:
+      - 'Ref A: 06AC2BB714854391AB2463BD25CB0D68 Ref B: YTO01EDGE0515 Ref C: 2023-02-01T16:48:44Z'
+      X-Tfs-Processid:
+      - b4314765-47f6-4a4e-aa52-8525ff053667
+      X-Tfs-Session:
+      - 88152494-f8a1-4314-b2eb-4a210fb00d73
+      X-Vss-E2eid:
+      - 88152494-f8a1-4314-b2eb-4a210fb00d73
+      X-Vss-Senderdeploymentid:
+      - a5ca35eb-148e-4ccd-bbb3-d31576d75958
+      X-Vss-Userdata:
+      - 473dec3e-03d7-6147-b106-0b0a7f766a92:idan.varsano@sourcegraph.com
+    status: 200 OK
+    code: 200
+    duration: ""
